+++ app/watcher/providers/docker/registries/dockerHub.test.js
@@ -37,3 +37,14 @@ test('authenticate should perform authenticate request', async () => {
     }));
     expect(dockerHub.authenticate({ login: 'login', password: 'password' })).resolves.toEqual('token');
 });
+
+test('base64Decode should decode credentials', () => {
+    expect(dockerHub.base64Decode('dXNlcm5hbWU6cGFzc3dvcmQ=')).toEqual({
+        login: 'username',
+        password: 'password',
+    });
+});
+
+test('base64Decode should throw exception when encoded credentials are wrong', () => {
+    expect(() => { dockerHub.base64Decode('wrong'); }).toThrow('Error when trying to get the login/password from the Base64 String (The Base64 decoded auth does not match with username:password pattern)');
+});